fixes on direct method

+++ SSA.h
@@ -18,10 +18,10 @@ protected:
     Utils* ut;
     DependencyGraph* dg;
     Model* model;
-    float* propArray; //propensity array
+    double* propArray; //propensity array
     double simulTime; //simulation time
     int* specQuantity; //species quantity
-    float totalPropensity; //sum of all the propensities
+    double totalPropensity; //sum of all the propensities
     //functions
     virtual void initialize(string filename) = 0; //read the model and initialize data structures
     virtual void calcPropensity() = 0; //calculate the propensity function for each reaction
